+++ src/command.nim
@@ -28,7 +28,7 @@ proc init*(conf: ProjectConfig) {.raises: [].} =
   # Initialise the correct volume-only container based on configured kind
   case conf.dbConf.kind
   of MySQL:
-    echo("Initialising " & $conf.dbConf.kind & " volume-only container..")
+    echo("Initialising " & $conf.dbConf.kind & " volume-only container...")
     shouldRunCommand = true
     command = join([
       "docker run -d -v",
@@ -40,7 +40,7 @@ proc init*(conf: ProjectConfig) {.raises: [].} =
       conf.dbConf.getImageName()
     ], " ")
   of PostgreSQL:
-    echo("Initialising " & $conf.dbConf.kind & " volume-only container..")
+    echo("Initialising " & $conf.dbConf.kind & " volume-only container...")
     shouldRunCommand = true
     command = join([
       "docker run -d -v",
@@ -54,12 +54,21 @@ proc init*(conf: ProjectConfig) {.raises: [].} =
       "/bin/echo",
       conf.dbConf.getImageName()
     ], " ")
+  of MongoDB:
+    echo("Initialising " & $conf.dbConf.kind & " volume-only container...")
+    shouldRunCommand = true
+    command = join([
+      "docker run -d -v",
+      conf.dbConf.getVolumePath(),
+      "--name",
+      conf.data,
+      "--entrypoint",
+      "/bin/echo",
+      conf.dbConf.getImageName()
+    ], " ")
   of None:
     echo("No database requested. If you change this in the future, you will need to reinitialise your dup project")
     shouldRunCommand = false
-  else:
-    echo("Error: Invalid database type specified in config")
-    quit(252)
   # Now check if we should build the state file, and run our command
   try:
     if shouldRunCommand == true:
@@ -101,7 +110,10 @@ proc up*(conf: ProjectConfig) {.raises: [].} =
   of PostgreSQL:
     startPostgres(conf.name, conf.dbConf.name, conf.dbConf.username, conf.dbConf.password)
     startWeb(conf.name, conf.port, conf.volume, conf.envVars, true)
-  else:
+  of MongoDB:
+    startMongo(conf)
+    startWeb(conf.name, conf.port, conf.volume, conf.envVars, true)
+  of None:
     # Start only the web container if we've got an invalid type (or None)
     startWeb(conf.name, conf.port, conf.volume, conf.envVars, false)
   quit(0)
@@ -187,6 +199,19 @@ proc sql*(conf: ProjectConfig) =
   of PostgreSQL:
     discard execCmd("docker exec -it -u postgres " & conf.db & " psql")
     quit(0)
+  of MongoDB:
+    discard execCmd("docker exec -it " & conf.db & " mongo")
+    quit(0)
   of None:
     echo("Error: No database configured for this project")
     quit(251)
+
+proc logs*(conf: ProjectConfig, args: Table[string, docopt.Value]) {.raises: [].} =
+  if args.getOrDefault("web"):
+    discard execCmd("docker logs -f " & conf.web)
+    quit(0)
+  if args.getOrDefault("db"):
+    discard execCmd("docker logs -f " & conf.db)
+    quit(0)
+  echo("Error: You must specify which container: 'dup sql web' or 'dup sql db'")
+  quit(250)