+++ CHANGES.md
@@ -1,11 +1,15 @@
 # Changelog
 
-## V1.0.0
+## V1.0.0-RC1
 
 - Massive refactoring of the core code. Properly setup in modules, properly typed with full compile-time checking and checked exception types.
 - Supports easy addition of new database types now
 - Code is now more robust, with proper error checking, and is now easier to test
 - Now respects `VIRTUAL_HOST` correctly, setting the hostname to the defined `env` value (if not specified, it still defaults to `<name>.docker`), resolving [issue #10](/girvo/dup/issues/10)
+- Database now supports `mongo` as a type parameter, with data-volume container handling for persistence
+ - This does not take any kind of config currently, merely: `"type": "mongo"`
+- Added `dup logs [web|db]` command
+ - Runs in "follow" mode, press Ctrl-C to stop viewing logs
 
 ## v0.4.0